# openapi.yaml
openapi: 3.1.0
info:
  title: Synthetic Data Generation Service
  version: 1.0.0
  description: |
    REST API для сервиса генерации конфиденциальных синтетических табличных данных
    с применением дифференциальной приватности (DP-CTGAN).

    ## Основной сценарий использования

    1. Загрузить датасет — `POST /datasets`
    2. Загрузить конфигурацию — `POST /configs`
    3. Запустить пайплайн — `POST /runs`
    4. Опрашивать статус — `GET /runs/{run_id}` (или дождаться webhook)
    5. Скачать синтетику — `GET /runs/{run_id}/synthetic`
    6. Изучить отчёт — `GET /runs/{run_id}/report`

    ## Асинхронное выполнение

    Обучение DP-CTGAN занимает несколько минут. Все запуски пайплайна
    (`POST /runs`) возвращают `run_id` немедленно со статусом `queued`.
    Итоговый статус — `completed` | `failed` | `cancelled`.

    ## Авторизация

    Все запросы требуют заголовок `Authorization: Bearer <token>`.
  contact:
    name: Сергей Рогачев
    email: s.rogachev@example.com
  license:
    name: MIT
servers:
- url: http://localhost:8000/api/v1
  description: Локальная разработка
- url: https://synth.internal/api/v1
  description: Внутренний production-стенд
...
security:
- BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          examples:
          - NOT_FOUND
        message:
          type: string
          examples:
          - Запуск с указанным ID не найден
        details:
          type: object
          additionalProperties: true
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          examples:
          - 42
        page:
          type: integer
          examples:
          - 1
        per_page:
          type: integer
          examples:
          - 20
        pages:
          type: integer
          examples:
          - 3
    RunStatus:
      type: string
      enum:
      - queued
      - running
      - completed
      - failed
      - cancelled
      description: |
        - `queued` — задача поставлена в очередь, обучение ещё не началось
        - `running` — пайплайн выполняется
        - `completed` — завершён успешно (вердикт может быть PASS, FAIL или PARTIAL)
        - `failed` — завершён с ошибкой
        - `cancelled` — отменён пользователем
    RunCreate:
      type: object
      required:
      - dataset_name
      - config_name
      properties:
        dataset_name:
          type: string
          description: Имя ранее загруженного датасета
          examples:
          - adult_census
        config_name:
          type: string
          description: Имя ранее загруженной конфигурации
          examples:
          - adult_epsilon3
        save_model:
          type: boolean
          default: false
          description: Сохранить обученную модель в model registry после завершения
        webhook_url:
          type:
          - string
          - 'null'
          format: uri
          description: URL для POST-уведомления при завершении запуска
          examples:
          - https://my-service.internal/hooks/synth-done
        n_synth_rows:
          type:
          - integer
          - 'null'
          minimum: 1
          description: Переопределить количество генерируемых строк (по умолчанию из конфига)
          examples:
          - 10000
    RunSummary:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
          examples:
          - 3fa85f64-5717-4562-b3fc-2c963f66afa6
        status:
          $ref: '#/components/schemas/RunStatus'
        verdict:
          type:
          - string
          - 'null'
          enum:
          - PASS
          - FAIL
          - PARTIAL
          - null
        dataset_name:
          type: string
          examples:
          - adult_census
        config_name:
          type: string
          examples:
          - adult_epsilon3
        created_at:
          type: string
          format: date-time
        started_at:
          type:
          - string
          - 'null'
          format: date-time
        finished_at:
          type:
          - string
          - 'null'
          format: date-time
        duration_sec:
          type:
          - number
          - 'null'
          examples:
          - 183.4
    RunDetail:
      allOf:
      - $ref: '#/components/schemas/RunSummary'
      - type: object
        properties:
          model_id:
            type:
            - string
            - 'null'
            description: ID модели в model registry (если save_model=true)
          synth_rows:
            type:
            - integer
            - 'null'
            description: Количество сгенерированных строк
          config_snapshot:
            type: object
            description: Снимок конфигурации на момент запуска
            additionalProperties: true
          error_message:
            type:
            - string
            - 'null'
            description: Сообщение об ошибке при status=failed
    RunReport:
      type: object
      description: Полный JSON-отчёт, аналогичный файлу из reporter/reports/
      properties:
        report_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        generated_at:
          type: string
          format: date-time
        verdict:
          $ref: '#/components/schemas/Verdict'
        data_processing:
          $ref: '#/components/schemas/DataProcessingReport'
        generator:
          $ref: '#/components/schemas/GeneratorReport'
        utility:
          $ref: '#/components/schemas/UtilityReport'
        privacy:
          $ref: '#/components/schemas/PrivacyReport'
    Verdict:
      type: object
      properties:
        overall:
          type: string
          enum:
          - PASS
          - FAIL
          - PARTIAL
        utility_ok:
          type:
          - boolean
          - 'null'
        privacy_ok:
          type:
          - boolean
          - 'null'
        dp_ok:
          type:
          - boolean
          - 'null'
        issues:
          type: array
          items:
            type: string
          examples:
          - - Utility Loss 0.320 > порога 0.150
            - MIA AUC 0.612 > порога 0.550 (риск утечки)
    DataProcessingReport:
      type: object
      properties:
        minimization:
          type: object
          properties:
            removed_direct_identifiers:
              type: array
              items:
                type: string
            removed_high_cardinality:
              type: array
              items:
                type: string
            columns_before:
              type: integer
            columns_after:
              type: integer
            rows_unchanged:
              type: boolean
    GeneratorReport:
      type: object
      properties:
        synthesizer:
          type: string
          examples:
          - smartnoise_dpctgan
        status:
          type: string
          examples:
          - fitted
        dp_config:
          type: object
          properties:
            epsilon_initial:
              type: number
              examples:
              - 3.0
            delta:
              type: number
              examples:
              - 4.47e-05
            sigma:
              type: number
              examples:
              - 5.0
            is_dp_enabled:
              type: boolean
        dp_spent:
          type: object
          properties:
            spent_epsilon_final:
              type:
              - number
              - 'null'
              examples:
              - 2.49
            epochs_requested:
              type: integer
              examples:
              - 300
            epochs_completed:
              type: integer
              examples:
              - 287
    UtilityReport:
      type: object
      properties:
        ml_efficacy:
          type: object
          properties:
            trtr:
              type: object
              properties:
                f1_weighted:
                  type: number
                  examples:
                  - 0.84
            tstr:
              type: object
              properties:
                f1_weighted:
                  type: number
                  examples:
                  - 0.71
            utility_loss:
              type: object
              properties:
                value:
                  type: number
                  examples:
                  - 0.13
        statistical:
          type: object
          properties:
            summary:
              type: object
              properties:
                mean_jsd:
                  type: number
                  examples:
                  - 0.14
                mean_tvd:
                  type: number
                  examples:
                  - 0.21
        correlations:
          type: object
          properties:
            pearson_corr_mae:
              type:
              - number
              - 'null'
              examples:
              - 0.05
...
    SampleRequest:
      type: object
      required:
      - n_rows
      properties:
        n_rows:
          type: integer
          minimum: 1
          maximum: 1000000
          examples:
          - 5000
        output_format:
          type: string
          enum:
          - csv
          - json
          - parquet
          default: csv
...
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Уникальный идентификатор запуска
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Невалидный запрос (ошибка валидации входных данных)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: VALIDATION_ERROR
                message: n_rows должен быть > 0
    Unauthorized:
      description: Отсутствует или невалиден токен авторизации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: UNAUTHORIZED
                message: Bearer token обязателен
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: NOT_FOUND
                message: Запуск не найден
...
    UnprocessableEntity:
      description: Семантически невалидный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
  /runs:
    get:
      tags:
      - runs
      summary: Список запусков
      description: Возвращает список всех запусков пайплайна с пагинацией. Можно фильтровать по статусу и вердикту.
      operationId: listRuns
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/RunStatus'
        description: Фильтр по статусу
      - name: verdict
        in: query
        schema:
          type: string
          enum:
          - PASS
          - FAIL
          - PARTIAL
        description: Фильтр по вердикту (только для завершённых)
      - name: dataset_name
        in: query
        schema:
          type: string
        description: Фильтр по датасету
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RunSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - runs
      summary: Запустить пайплайн
      description: |
        Создаёт новый запуск пайплайна и ставит его в очередь.
        Возвращает `run_id` немедленно — обучение выполняется асинхронно.
        Опрашивайте `GET /runs/{run_id}` для отслеживания прогресса,
        или укажите `webhook_url` для push-уведомления при завершении.
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreate'
            examples:
              default:
                value:
                  dataset_name: adult_census
                  config_name: adult_epsilon3
                  save_model: true
                  webhook_url: https://my-service.internal/hooks/synth-done
      responses:
        '202':
          description: Запуск поставлен в очередь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSummary'
              examples:
                default:
                  value:
                    run_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    status: queued
                    verdict: null
                    dataset_name: adult_census
                    config_name: adult_epsilon3
                    created_at: '2026-04-09T12:00:00Z'
                    started_at: null
                    finished_at: null
                    duration_sec: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Датасет или конфиг не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                default:
                  value:
                    code: NOT_FOUND
                    message: Датасет 'adult_census' не найден. Сначала загрузите его через POST /datasets
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /runs/{run_id}:
    get:
      tags:
      - runs
      summary: Статус и детали запуска
      operationId: getRun
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - runs
      summary: Отменить или удалить запуск
      description: |
        Если статус `queued` или `running` — отменяет выполнение.
        Если `completed` или `failed` — удаляет запись и связанные файлы (синтетику, отчёт).
        Модель в registry при этом **не** удаляется.
      operationId: deleteRun
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '204':
          description: Успешно удалён / отменён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /runs/{run_id}/report:
    get:
      tags:
      - runs
      summary: Получить отчёт о запуске
      description: Возвращает полный JSON-отчёт с вердиктом, метриками полезности и приватности. Доступен только для завершённых
        запусков.
      operationId: getRunReport
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Запуск ещё не завершён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                default:
                  value:
                    code: RUN_NOT_FINISHED
                    message: 'Отчёт недоступен — запуск ещё выполняется (status: running)'
  /runs/{run_id}/logs:
    get:
      tags:
      - runs
      summary: Получить лог запуска
      operationId: getRunLogs
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: tail
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 10000
        description: Вернуть только последние N строк (по умолчанию — весь лог)
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              examples:
                default:
                  value: |
                    [INFO] Загрузка конфига: configs/adult.yaml
                    [INFO] Загружено строк: 48842
                    [INFO] Обучение генератора на real_train...
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /runs/{run_id}/synthetic:
    get:
      tags:
      - synthetic
      summary: Скачать синтетику запуска
      description: Возвращает сгенерированный датасет в указанном формате. Доступен только для завершённых запусков.
      operationId: getRunSynthetic
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: format
        in: query
        schema:
          type: string
          enum:
          - csv
          - json
          - parquet
          default: csv
      responses:
        '200':
          description: Файл синтетических данных
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              examples:
                default:
                  value: attachment; filename="adult_census__dpctgan__20260409.csv"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Запуск ещё не завершён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - synthetic
      summary: Догенерировать строки из модели запуска
      description: |
        Генерирует дополнительные строки с использованием модели этого запуска
        без повторного обучения. Требует `save_model: true` при создании запуска.
        Не тратит дополнительный ε-бюджет.
      operationId: generateMoreSynthetic
      parameters:
      - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SampleRequest'
            examples:
              default:
                value:
                  n_rows: 5000
                  output_format: csv
      responses:
        '200':
          description: Файл дополнительной синтетики
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Модель не была сохранена для этого запуска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                default:
                  value:
                    code: MODEL_NOT_SAVED
                    message: 'Модель не сохранена. Повторите запуск с save_model: true'
...